fixed opacity issue in timeline

+++ lvgl-runtime/common/src/flow.cpp
@@ -116,7 +116,7 @@ struct WidgetTimeline {
 	int16_t y;
 	int16_t width;
 	int16_t height;
-    int16_t opacity;
+    float opacity;
     int16_t scale;
     int16_t rotate;
 
@@ -134,7 +134,7 @@ void addTimelineKeyframe(
     int16_t y, uint8_t yEasingFunc,
     int16_t width, uint8_t widthEasingFunc,
     int16_t height, uint8_t heightEasingFunc,
-    int16_t opacity, uint8_t opacityEasingFunc,
+    float opacity, uint8_t opacityEasingFunc,
     int16_t scale, uint8_t scaleEasingFunc,
     int16_t rotate, uint8_t rotateEasingFunc,
     int32_t cp1x, int32_t cp1y, int32_t cp2x, int32_t cp2y
